# Filter undesirable fences (PR against develop branch)