--- a/debugging/kernel-logger/kp_kernel_logger.cpp
+++ b/debugging/kernel-logger/kp_kernel_logger.cpp
@@ -121,14 +121,18 @@ extern "C" void kokkosp_end_parallel_reduce(const uint64_t kID) {
   printf("KokkosP: Execution of kernel %llu is completed.\n", kID);
 }
 
-extern "C" void kokkosp_begin_fence(const char* name, const uint32_t devID,
-                                    uint64_t* kID) {
-  *kID = uniqID++;
+extern "C" void kokkosp_begin_fence(const char* name, const uint32_t devID, uint64_t* kID) {
+		
+	 if (std::strstr(name, "Kokkos Profile Tool Fence"))
+           {
+              *kID = std::numeric_limits<uint64_t>::max();
+           }
+        else
+          {
+          *kID = uniqID++;
 
-  printf(
-      "KokkosP: Executing fence on device %d with unique execution identifier "
-      "%llu\n",
-      devID, *kID);
+        printf("KokkosP: Executing fence on device %d with unique execution identifier %llu\n",
+                devID, *kID);
 
   int level = kokkosp_print_region_stack();
   kokkosp_print_region_stack_indent(level);