@vlkale 2022-12-08
Note that this PR still needs to handle the other callbacks and other tools, e.g., the parallel_for callback in the roctx-connector. This is what is in progress.